test: colocate tests in `core`
