+++ core/diff/streamDiff.test.ts
