+++ core/indexing/chunk/code.test.ts
@@ -1,5 +1,5 @@
-import { ChunkWithoutID } from "../../../index.js";
-import { codeChunker } from "../../../indexing/chunk/code.js";
+import { ChunkWithoutID } from "../..";
+import { codeChunker } from "./code";
 
 async function genToArr<T>(generator: AsyncGenerator<T>): Promise<T[]> {
   const result: T[] = [];
@@ -29,7 +29,7 @@ describe.skip("codeChunker", () => {
   test("should capture small class and function from large python file", async () => {
     const extraLine = "# This is a comment";
     const myClass = "class MyClass:\n    def __init__(self):\n        pass";
-    const myFunction = 'def my_function():\n    return "Hello, World!"';
+    const myFunction = "def my_function():\n    return \"Hello, World!\"";
 
     const file =
       Array(100).fill(extraLine).join("\n") +
@@ -66,7 +66,7 @@ describe.skip("codeChunker", () => {
       chunks[0].startsWith("class MyClass:\n    def method1():\n        ..."),
     ).toBe(true);
     // The extra spaces seem to be a bug with tree-sitter-python
-    expect(chunks).toContain('def method1():\n        return "Hello, 1!"');
-    expect(chunks).toContain('def method20():\n        return "Hello, 20!"');
+    expect(chunks).toContain("def method1():\n        return \"Hello, 1!\"");
+    expect(chunks).toContain("def method20():\n        return \"Hello, 20!\"");
   });
 });
