+++ core/test/context/providers/index.skip.ts
@@ -8,7 +8,7 @@ import { ConfigHandler } from "../../../config/ConfigHandler";
 import { contextProviderClassFromName } from "../../../context/providers";
 import { ControlPlaneClient } from "../../../control-plane/client";
 import FileSystemIde from "../../../util/filesystem";
-import { TEST_DIR } from "../../testUtils/testDir";
+import { TEST_DIR } from "../../util/testDir";
 
 const CONTEXT_PROVIDERS_TO_TEST: ContextProviderWithParams[] = [
   { name: "diff", params: {} },
