--- core/test/context/retrieval/RetrievalPipeline.test.ts
@@ -1,24 +1,0 @@
-import { IRetrievalPipeline } from "../../../context/retrieval/pipelines/BaseRetrievalPipeline.js";
-
-function testRetrievalPipeline(pipeline: IRetrievalPipeline) {
-  test("should successfully run");
-}
-
-describe.skip("RetrievalPipeline", () => {
-  test.todo("should successfully run");
-});
-
-// describe("Retrieval Pipelines", () => {
-//   const ide = new FileSystemIde();
-//   const options: RetrievalPipelineOptions;
-
-//   test("NoRerankerRetrievalPipeline", () => {
-//     const pipeline = new NoRerankerRetrievalPipeline(options);
-//     testRetrievalPipeline(pipeline);
-//   });
-
-//   test("NoRerankerRetrievalPipeline", () => {
-//     const pipeline = new NoRerankerRetrievalPipeline(options);
-//     testRetrievalPipeline(pipeline);
-//   });
-// });
