+++ core/test/llm.test.ts
@@ -1,8 +1,8 @@
 import * as dotenv from "dotenv";
 
-import { CompletionOptions } from "../index.js";
-import { BaseLLM } from "../llm/index.js";
-import OpenAI from "../llm/llms/OpenAI.js";
+import { CompletionOptions } from "../";
+import { BaseLLM } from "../llm/";
+import OpenAI from "../llm/llms/OpenAI";
 
 // jest.setTimeout(100_000);
