--- core/test/util/dedentAndGetCommonWhitespace.test.ts
@@ -1,76 +1,0 @@
-// File generated by Continue
-import { dedentAndGetCommonWhitespace } from "../../util/index.js";
-
-describe("dedentAndGetCommonWhitespace", () => {
-  let originalString: string;
-
-  beforeEach(() => {
-    // Setup any global variables or states if needed
-    originalString = "    line1\n    line2\n    line3";
-  });
-
-  afterEach(() => {
-    // Tear down any changes to global variables or states if needed
-    originalString = "";
-  });
-
-  test("should dedent and return common whitespace for a simple case", () => {
-    const input = "    line1\n    line2\n    line3";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual(["line1\nline2\nline3", "    "]);
-  });
-
-  test("should handle empty string", () => {
-    const input = "";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual(["", ""]);
-  });
-
-  test("should handle string with only whitespace", () => {
-    const input = "    ";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual(["", ""]);
-  });
-
-  test("should handle string with mixed whitespace and content", () => {
-    const input = "    line1\n  line2\n    line3";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual(["  line1\nline2\n  line3", "  "]);
-  });
-
-  test("should handle string with no common leading whitespace", () => {
-    const input = "line1\n  line2\n    line3";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual([input, ""]);
-  });
-
-  test("should handle string with empty lines", () => {
-    const input = "    line1\n\n    line3";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual(["line1\n\nline3", "    "]);
-  });
-
-  test("should handle string with only empty lines", () => {
-    const input = "\n\n";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual(["\n\n", ""]);
-  });
-
-  test("should handle string with tabs as whitespace", () => {
-    const input = "\tline1\n\tline2\n\tline3";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual(["line1\nline2\nline3", "\t"]);
-  });
-
-  test("should handle string with mixed tabs and spaces", () => {
-    const input = "\t    line1\n\t    line2\n\t    line3";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual(["line1\nline2\nline3", "\t    "]);
-  });
-
-  test("should handle string with different leading whitespace lengths", () => {
-    const input = "    line1\n  line2\n      line3";
-    const output = dedentAndGetCommonWhitespace(input);
-    expect(output).toEqual(["  line1\nline2\n    line3", "  "]);
-  });
-});
