+++ core/test/util/fixtures.ts
@@ -1,7 +1,7 @@
-import { ConfigHandler } from "../config/ConfigHandler";
-import { ControlPlaneClient } from "../control-plane/client";
-import FileSystemIde from "../util/filesystem";
-import { TEST_DIR } from "./testUtils/testDir";
+import { ConfigHandler } from "../../config/ConfigHandler";
+import { ControlPlaneClient } from "../../control-plane/client";
+import FileSystemIde from "../../util/filesystem";
+import { TEST_DIR } from "./testDir";
 
 export const testIde = new FileSystemIde(TEST_DIR);
