--- core/test/util/getLastNPathParts.test.ts
@@ -1,51 +1,0 @@
-// File generated by Continue
-import { getLastNPathParts } from "../../util/index.js";
-
-describe("getLastNPathParts", () => {
-  const SEP_REGEX = /[\\/]+/;
-
-  beforeAll(() => {
-    // Setup code if needed
-  });
-
-  afterAll(() => {
-    // Teardown code if needed
-  });
-
-  test("returns the last N parts of a filepath with forward slashes", () => {
-    const filepath = "home/user/documents/project/file.txt";
-    expect(getLastNPathParts(filepath, 2)).toBe("project/file.txt");
-  });
-
-  test("returns the last N parts of a filepath with backward slashes", () => {
-    const filepath = "C:\\home\\user\\documents\\project\\file.txt";
-    expect(getLastNPathParts(filepath, 3)).toBe("documents/project/file.txt");
-  });
-
-  test("returns the last part if N is 1", () => {
-    const filepath = "/home/user/documents/project/file.txt";
-    expect(getLastNPathParts(filepath, 1)).toBe("file.txt");
-  });
-
-  test("returns the entire path if N is greater than the number of parts", () => {
-    const filepath = "home/user/documents/project/file.txt";
-    expect(getLastNPathParts(filepath, 10)).toBe(
-      "home/user/documents/project/file.txt",
-    );
-  });
-
-  test("returns an empty string if N is 0", () => {
-    const filepath = "home/user/documents/project/file.txt";
-    expect(getLastNPathParts(filepath, 0)).toBe("");
-  });
-
-  test("handles paths with mixed forward and backward slashes", () => {
-    const filepath = "home\\user/documents\\project/file.txt";
-    expect(getLastNPathParts(filepath, 3)).toBe("documents/project/file.txt");
-  });
-
-  test("handles edge case with empty filepath", () => {
-    const filepath = "";
-    expect(getLastNPathParts(filepath, 2)).toBe("");
-  });
-});
