--- core/test/util/lcs.test.ts
@@ -1,40 +1,0 @@
-// This file is generated by Continue
-import { longestCommonSubsequence } from "../../util/lcs.js";
-
-describe("longestCommonSubsequence", () => {
-  beforeAll(() => {
-    // Setup if needed
-  });
-
-  afterAll(() => {
-    // Teardown if needed
-  });
-
-  test("should return the longest common subsequence for non-empty strings", () => {
-    expect(longestCommonSubsequence("abcde", "ace")).toBe("ace");
-    expect(longestCommonSubsequence("abc", "abc")).toBe("abc");
-    expect(longestCommonSubsequence("abc", "def")).toBe("");
-  });
-
-  test("should handle cases where one or both strings are empty", () => {
-    expect(longestCommonSubsequence("", "abc")).toBe("");
-    expect(longestCommonSubsequence("abc", "")).toBe("");
-    expect(longestCommonSubsequence("", "")).toBe("");
-  });
-
-  test("should handle cases with no common subsequence", () => {
-    expect(longestCommonSubsequence("abc", "def")).toBe("");
-    expect(longestCommonSubsequence("xyz", "abc")).toBe("");
-  });
-
-  test("should handle cases with special characters", () => {
-    expect(longestCommonSubsequence("a!@#b$c", "!@#$")).toBe("!@#$");
-    expect(longestCommonSubsequence("a!@#b$c", "xyz")).toBe("");
-  });
-
-  test("should handle long strings efficiently", () => {
-    const str1 = "a".repeat(1000) + "b".repeat(1000);
-    const str2 = "a".repeat(1000) + "c".repeat(1000);
-    expect(longestCommonSubsequence(str1, str2)).toBe("a".repeat(1000));
-  });
-});
